Lots of Changes
Added full update functionality for customers as well as delete functionality for customers, vendors, and products. The database still needs to be updated to allow for cascading deletes. Additionally the cart delete function works, but you have to directly do a get request in the browser instead of doing it through the website since the cart does not have full functionality yet.